requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.3.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
jinja2>=3.1.0
aiofiles>=23.0.0
websockets>=12.0
